api/requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cron_descriptor==1.4.0
Django==4.2.9
django-celery-beat==2.5.0
django-dotenv==1.4.2
django-timezone-field==6.1.0
kombu==5.3.4
mpmath==1.3.0
mysqlclient==2.2.1
numpy==1.26.3
prompt-toolkit==3.0.43
PyMySQL==1.1.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tzdata==2023.4
vine==5.1.0
wcwidth==0.2.12
